+++ sklearn/feature_extraction/text.py
@@ -4,6 +4,7 @@
 #          Lars Buitinck <L.J.Buitinck@uva.nl>
 #          Robert Layton <robertlayton@gmail.com>
 #          Jochen Wersdörfer <jochen@wersdoerfer.de>
+#          Roman Sinayev <roman.sinayev@gmail.com>
 #
 # License: BSD Style.
 """
@@ -13,7 +14,7 @@ build feature vectors from text documents.
 from __future__ import unicode_literals
 
 import array
-from collections import Mapping
+from collections import Mapping, defaultdict
 import numbers
 from operator import itemgetter
 import re
@@ -26,7 +27,6 @@ import scipy.sparse as sp
 from ..base import BaseEstimator, TransformerMixin
 from ..externals.six.moves import xrange
 from ..preprocessing import normalize
-from ..utils.fixes import Counter
 from .hashing import FeatureHasher
 from .stop_words import ENGLISH_STOP_WORDS
 from sklearn.externals import six
@@ -149,12 +149,12 @@ class VectorizerMixin(object):
         Tokenize text_document into a sequence of character n-grams
         excluding any whitespace (operating only inside word boundaries)"""
         # normalize white spaces
-        text_document = self._white_spaces.sub(" ", text_document)
+        text_document = self._white_spaces.sub(u" ", text_document)
 
         min_n, max_n = self.ngram_range
         ngrams = []
         for w in text_document.split():
-            w = " " + w + " "
+            w = ' ' + w + ' '
             w_len = len(w)
             for n in xrange(min_n, max_n + 1):
                 offset = 0
@@ -448,7 +448,7 @@ class CountVectorizer(BaseEstimator, VectorizerMixin):
 
     If you do not provide an a-priori dictionary and you do not use an analyzer
     that does some kind of feature selection then the number of features will
-    be equal to the vocabulary size found by analysing the data.
+    be equal to the vocabulary size found by analyzing the data.
 
     Parameters
     ----------
@@ -591,7 +591,15 @@ class CountVectorizer(BaseEstimator, VectorizerMixin):
         self.stop_words = stop_words
         self.max_df = max_df
         self.min_df = min_df
+        if max_df < 0 or min_df < 0:
+            raise ValueError("negative value for max_df of min_df")
         self.max_features = max_features
+        if not any((
+                isinstance(max_features, numbers.Integral),
+                max_features is None,
+                max_features > 0)):
+            raise ValueError(
+                "max_features is neither a positive integer nor None")
         self.ngram_range = ngram_range
         if vocabulary is not None:
             if not isinstance(vocabulary, Mapping):
@@ -599,17 +607,21 @@ class CountVectorizer(BaseEstimator, VectorizerMixin):
             if not vocabulary:
                 raise ValueError("empty vocabulary passed to fit")
             self.fixed_vocabulary = True
-            self.vocabulary_ = vocabulary
+            self.vocabulary_ = dict(vocabulary)
         else:
             self.fixed_vocabulary = False
         self.binary = binary
         self.dtype = dtype
 
-    def _term_counts_to_matrix(self, n_doc, i_indices, j_indices, values):
+    def _term_counts_to_matrix(self, n_doc, i_indices, j_indices, values,
+                               n_features=None):
         """Construct COO matrix from indices and values.
 
         i_indices and j_indices should be constructed with _make_int_array.
+
         """
+        n_features = n_features or len(self.vocabulary_)
+
         # array("i") corresponds to np.intc, which is also what scipy.sparse
         # wants for indices, so they won't be copied by the coo_matrix ctor.
         # The length check works around a bug in old NumPy versions:
@@ -618,7 +630,6 @@ class CountVectorizer(BaseEstimator, VectorizerMixin):
             i_indices = np.frombuffer(i_indices, dtype=np.intc)
         if len(j_indices) > 0:
             j_indices = np.frombuffer(j_indices, dtype=np.intc)
-
         if self.dtype == np.intc and len(values) > 0:
             values = np.frombuffer(values, dtype=np.intc)
         else:
@@ -626,12 +637,73 @@ class CountVectorizer(BaseEstimator, VectorizerMixin):
             # array.array.
             values = np.asarray(values, dtype=self.dtype)
 
-        shape = (n_doc, max(six.itervalues(self.vocabulary_)) + 1)
+        shape = (n_doc, n_features)
         spmatrix = sp.coo_matrix((values, (i_indices, j_indices)),
                                  shape=shape, dtype=self.dtype)
-        if self.binary:
-            spmatrix.data.fill(1)
         return spmatrix
+
+    def _sort_features_and_matrix(self, cscmatrix, feature_to_pos):
+        '''sort dict by keys and assign values to the sorted key index'''
+        sorted_by_name = {}
+        sorted_features = sorted(feature_to_pos)
+        new_positions = []
+
+        for i, k in enumerate(sorted_features):
+            sorted_by_name[k] = i
+            new_positions.append(feature_to_pos[k])
+
+        return cscmatrix[:, new_positions], sorted_by_name
+
+    def _remove_highandlow(self, cscmatrix, feature_to_pos, high, low):
+        """Remove too rare or too common features.
+
+        Prune features that are non zero in more samples than high or less
+        documents than low.
+
+        This does not prune samples with zero features.
+
+        """
+        kept_indices = []
+        removed_indices = set()
+        for colptr in xrange(len(cscmatrix.indptr) - 1):
+            len_slice = (cscmatrix.indptr[colptr + 1]
+                       - cscmatrix.indptr[colptr])
+            if len_slice <= high and len_slice >= low:
+                kept_indices.append(colptr)
+            else:
+                removed_indices.add(colptr)
+
+        s_kept_indices = set(kept_indices)
+        new_mapping = dict((v, i) for i, v in enumerate(kept_indices))
+        feature_to_pos = dict((k, new_mapping[v])
+                              for k, v in six.iteritems(feature_to_pos)
+                              if v in s_kept_indices)
+
+        return cscmatrix[:, kept_indices], feature_to_pos, removed_indices
+
+    def _get_kept_features(self, csc_m, max_features):
+        '''helper method for _get_max_features to use less memory'''
+        feature_freqs = np.asarray(csc_m.sum(axis=0)).ravel()
+        to_keep = np.argsort(feature_freqs)[::-1][:max_features]
+        return to_keep
+
+    def _get_max_features(self, csc_m, feature_to_pos,
+                          stop_words_, max_features):
+        '''Remove maximum features using a sparse matrix.
+
+        Cut only the top max_features from the matrix and
+        feature_to_pos. Cut features are added to stop_words_.
+
+        '''
+        to_keep = self._get_kept_features(csc_m, max_features)
+        s_to_keep = set(to_keep)
+        new_feature_to_pos = {}
+        for k, v in six.iteritems(feature_to_pos):
+            if v in s_to_keep:
+                new_feature_to_pos[k] = v
+            else:
+                stop_words_.add(k)
+        return csc_m[:, np.sort(to_keep)], new_feature_to_pos, stop_words_
 
     def fit(self, raw_documents, y=None):
         """Learn a vocabulary dictionary of all tokens in the raw documents.
@@ -666,131 +738,86 @@ class CountVectorizer(BaseEstimator, VectorizerMixin):
         # fit_transform overridable without unwanted side effects in
         # TfidfVectorizer.
         fixed_vocab = self.fixed_vocabulary
-
-        if fixed_vocab:
-            vocab = self.vocabulary_
-            vocab_max_ind = max(six.itervalues(self.vocabulary_)) + 1
-        else:
-            vocab = {}
-            vocab_max_ind = 0
-
-        # Result of document conversion to term count arrays.
-        row_ind = _make_int_array()
-        col_ind = _make_int_array()
-        feature_values = _make_int_array()
-        term_counts = Counter()
-
-        # term counts across entire corpus (count each term maximum once per
-        # document)
-        document_counts = Counter()
-
         analyze = self.build_analyzer()
-
-        for n_doc, doc in enumerate(raw_documents):
-            term_count_current = Counter(analyze(doc))
-            term_counts.update(term_count_current)
-
-            if not fixed_vocab:
-                for term in six.iterkeys(term_count_current):
-                    if term not in vocab:
-                        vocab[term] = vocab_max_ind
-                        vocab_max_ind += 1
-
-            document_counts.update(six.iterkeys(term_count_current))
-
-            for term, count in six.iteritems(term_count_current):
-                if term in vocab:
-                    row_ind.append(n_doc)
-                    col_ind.append(vocab[term])
-                    feature_values.append(count)
-        n_doc += 1
-
+        max_df = self.max_df
+        min_df = self.min_df
+        max_features = self.max_features
+        binary = self.binary
+        i_indices = _make_int_array()
+        j_indices = _make_int_array()
+        values = _make_int_array()
+        # we create 3 arrays with i_indices corresponding to samples
+        # ordered in the order first encountered,
+        # j_indices corresponding to features also ordered in order
+        # first encountered.  Values corresponds to the value in that
+        # position.  We take advantage of the fact that values in
+        # the same position get implicitly added together when the COO
+        # matrix is constructed.
         if fixed_vocab:
-            # XXX max_df, min_df and max_features have no effect
-            # with a fixed vocabulary.
-            i_indices = row_ind
-            j_indices = col_ind
-            values = feature_values
+            feature_to_pos = self.vocabulary_
+            for i, doc in enumerate(raw_documents):
+                for feature in analyze(doc):
+                    try:
+                        j_indices.append(feature_to_pos[feature])
+                        i_indices.append(i)
+                        values.append(1)
+                    except KeyError:
+                        continue
+            n_doc = i + 1
         else:
-            max_features = self.max_features
-            max_df = self.max_df
-            min_df = self.min_df
+            j = 0
+            feature_to_count = defaultdict(int)
+            feature_to_pos = defaultdict(int)
+            for i, doc in enumerate(raw_documents):
+                for feature in analyze(doc):
+                    feature_to_count[feature] += 1
+                    if feature_to_count[feature] == 1:  # new feature
+                        feature_to_pos[feature] = j
+                        j += 1
+                    i_indices.append(i)
+                    j_indices.append(feature_to_pos[feature])
+                    values.append(1)
+            # assert j == len(feature_to_count)
+            del feature_to_count  # free memory
+            n_doc = i + 1
+            max_doc_count = (max_df if isinstance(max_df, numbers.Integral)
+                                    else int(round(max_df * n_doc)))
+            min_doc_count = (min_df if isinstance(min_df, numbers.Integral)
+                                    else int(round(min_df * n_doc)))
+            if max_doc_count < min_doc_count:
+                raise ValueError(
+                    "max_df corresponds to < documents than min_df")
 
-            max_doc_count = (max_df
-                             if isinstance(max_df, numbers.Integral)
-                             else max_df * n_doc)
-            min_doc_count = (min_df
-                             if isinstance(min_df, numbers.Integral)
-                             else min_df * n_doc)
+        n_features = len(feature_to_pos)
+        if n_features == 0:
+            raise ValueError("empty vocabulary;"
+                             " perhaps the document only contains stop words")
 
-            # filter out stop words: terms that occur in almost all documents
+        csc_m = self._term_counts_to_matrix(n_doc, i_indices, j_indices,
+                                            values, n_features).tocsc()
+        del i_indices, j_indices, values  # free memory
+        if binary:
+            csc_m.data.fill(1)
+        if not fixed_vocab:
+            csc_m, feature_to_pos = self._sort_features_and_matrix(csc_m,
+                                                               feature_to_pos)
+            stop_words_ = set()
+            # get rid of features between max_df and min_df
             if max_doc_count < n_doc or min_doc_count > 1:
-                stop_words = set(t for t, dc in six.iteritems(document_counts)
-                                 if not min_doc_count <= dc <= max_doc_count)
-            else:
-                stop_words = set()
+                csc_m, feature_to_pos, stop_words_ = \
+                                    self._remove_highandlow(
+                                    csc_m, feature_to_pos,
+                                    max_doc_count, min_doc_count)
+            # get rid of features that are not in the top max_features
+            # overall occurance wise
+            if max_features and max_features < csc_m.shape[1]:
+                csc_m, feature_to_pos, stop_words_ = \
+                    self._get_max_features(
+                        csc_m, feature_to_pos, stop_words_, max_features)
+            self.stop_words_ = stop_words_
+            self.vocabulary_ = feature_to_pos
 
-            # list the terms that should be part of the vocabulary
-            if max_features is None:
-                terms = set(term_counts) - stop_words
-            else:
-                # extract the most frequent terms for the vocabulary
-                terms = set()
-                for t, tc in term_counts.most_common():
-                    if t not in stop_words:
-                        terms.add(t)
-                    if len(terms) >= max_features:
-                        break
-
-            # store the learned stop words to make it easier to debug the value
-            # of max_df
-            self.stop_words_ = stop_words
-
-            # free memory
-            term_counts.clear()
-            document_counts.clear()
-
-            # store map from term name to feature integer index: we sort the
-            # terms to have reproducible outcome for the vocabulary structure:
-            # otherwise the mapping from feature name to indices might depend
-            # on the memory layout of the machine. Furthermore sorted terms
-            # might make it possible to perform binary search in the feature
-            # names array.
-            terms = sorted(terms)
-
-            # reorder term indices
-            reorder_indices = dict((vocab[term], i)
-                                   for i, term in enumerate(terms))
-            self.vocabulary_ = dict(((t, i) for i, t in enumerate(terms)))
-
-            # create term count arrays with new vocabulary structure
-            i_indices = _make_int_array()
-            j_indices = _make_int_array()
-            values = _make_int_array()
-            for i, col in enumerate(col_ind):
-                if col in reorder_indices:
-                    i_indices.append(row_ind[i])
-                    j_indices.append(reorder_indices[col_ind[i]])
-                    values.append(feature_values[i])
-
-            # free memory
-            del reorder_indices
-            del row_ind
-            del col_ind
-            del feature_values
-
-        if not vocab:
-            msg = "Empty vocabulary; "
-            if fixed_vocab:
-                msg += "%r passed to constructor." % vocab
-            else:
-                msg += "perhaps your documents contain stop words only?"
-            raise ValueError(msg)
-
-        # the term_counts and document_counts might be useful statistics, are
-        # we really sure want we want to drop them? They take some memory but
-        # can be useful for corpus introspection
-        return self._term_counts_to_matrix(n_doc, i_indices, j_indices, values)
+        return csc_m
 
     def transform(self, raw_documents):
         """Extract token counts out of raw text documents using the vocabulary
@@ -810,23 +837,28 @@ class CountVectorizer(BaseEstimator, VectorizerMixin):
 
         # raw_documents can be an iterable so we don't know its size in
         # advance
-
-        # result of document conversion to term count arrays
         i_indices = _make_int_array()
         j_indices = _make_int_array()
         values = _make_int_array()
-
+        binary = self.binary
         analyze = self.build_analyzer()
+        # use the same matrix-building strategy as fit_transform
         for n_doc, doc in enumerate(raw_documents):
-            term_counts = Counter(analyze(doc))
-
-            for term, count in six.iteritems(term_counts):
-                if term in self.vocabulary_:
+            for feature in analyze(doc):
+                try:
+                    j_indices.append(self.vocabulary_[feature])
                     i_indices.append(n_doc)
-                    j_indices.append(self.vocabulary_[term])
-                    values.append(count)
+                    values.append(1)
+                except KeyError:  # feature not in the vectorizer vocabulary
+                    pass
         n_doc += 1
-        return self._term_counts_to_matrix(n_doc, i_indices, j_indices, values)
+        n_features = len(self.vocabulary_)
+        m = self._term_counts_to_matrix(n_doc, i_indices, j_indices, values)
+
+        del i_indices, j_indices, values  # free memory
+        if binary:
+            m.data.fill(1)
+        return m.tocsc()  # keep the same sparse format as fit_transform
 
     def inverse_transform(self, X):
         """Return terms per document with nonzero entries in X.
@@ -840,7 +872,8 @@ class CountVectorizer(BaseEstimator, VectorizerMixin):
         X_inv : list of arrays, len = n_samples
             List of arrays of terms.
         """
-        if sp.isspmatrix_coo(X):  # COO matrix is not indexable
+        if sp.isspmatrix_coo(X) or sp.isspmatrix_csc(X):
+            # COO matrix is not indexable, CSC is slow for row manipulations
             X = X.tocsr()
         elif not sp.issparse(X):
             # We need to convert X to a matrix, so that the indexing